The-East-West-Alliance/StreamlitApp/requirements.txt
beautifulsoup4==4.11.1
joblib==1.1.0
matplotlib==3.5.1
nltk==3.7
numpy==1.20.3
pandas==1.3.0
Pillow==9.1.0
spacy==3.2.4
wordcloud==1.8.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz